Rename cpu -> vm
