+++ executor/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod cpu;
 pub mod elf;
+pub mod vm;
