+++ executor/src/vm/logs.rs
@@ -1,4 +1,4 @@
-use crate::cpu::instruction::decoding::Instruction;
+use crate::vm::instruction::decoding::Instruction;
 
 /// Log containing the executed instruction and the new value of the updated register
 pub struct Log {
